# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LASU Fleet Manager API",
        "description": "API documentation for the LASU fleet management system.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/owners/{ownerId}/maintenance-requests": {
            "get": {
                "tags": [
                    "Owner"
                ],
                "summary": "List maintenance requests for owner",
                "operationId": "637d050f59e709bcc209fbadc30fa168",
                "parameters": [
                    {
                        "name": "ownerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of maintenance requests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Owner"
                ],
                "summary": "Submit a maintenance request",
                "operationId": "3b7608f9afd94a034d22336a18c0aeed",
                "parameters": [
                    {
                        "name": "ownerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicleId",
                                    "issue",
                                    "priority"
                                ],
                                "properties": {
                                    "vehicleId": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "issue": {
                                        "type": "string",
                                        "example": "Brake failure"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ],
                                        "example": "high"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}